Put SSL verify_peer check in right place
I put this in the wrong place in 9996b7b. In the pre-9996b7b version of
the code, when verify_peer was false, no server certificate verification
was performed.

I had misread the previous code and thought that verify_peer being false
just meant that it would not check that the server certificate matched
the hostname, which is what the warning message suggested; I’ve fixed
the warning message here too.

Also, my mistake in 9996b7b did introduce a bunch of test failures, but
I didn’t notice them because on GitHub Actions these tests weren’t being
run due to some misbehaviour of the requires_connection test helper;
I’ll remove that helper in an upcoming commit.